Changed utf8 functions to take an output buffer and length, and return the necessary length like the bsd strl functions.
This version does not seem to paste or d&d foreign characters on
Win32, hope somebody can fix it.


git-svn-id: file:///fltk/svn/trunk@3700 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/fl_xpm.cxx
@@ -1,4 +1,4 @@
-// "$Id: fl_xpm.cxx,v 1.24 2004/07/07 05:11:03 spitzak Exp $"
+// "$Id: fl_xpm.cxx,v 1.25 2004/07/27 07:03:08 spitzak Exp $"
 // Copyright 1998-2004 by Bill Spitzak and others.
 //
 // This library is free software; you can redistribute it and/or
@@ -127,7 +127,7 @@ static char** read(char *name, int oneline = 0) {
   looks like the start of a .xpm file. This returns true if the 
   data contains "/\* XPM".
 */
-bool xpmFileImage::test(const unsigned char *data, size_t)
+bool xpmFileImage::test(const unsigned char *data, unsigned)
 {
   return (strstr((char*) data,"/* XPM") != 0);
 }
@@ -196,5 +196,5 @@ void xpmFileImage::read()
 }
 
 //
-// End of "$Id: fl_xpm.cxx,v 1.24 2004/07/07 05:11:03 spitzak Exp $"
+// End of "$Id: fl_xpm.cxx,v 1.25 2004/07/27 07:03:08 spitzak Exp $"
 //
